fix expression to work with new fixity of .+^ (fixes #27)
See https://github.com/conal/vector-space/commit/f5ef4fd .  Note that
we could make use of the new left-associativity to omit some
parentheses here, which looks nice, but then we would have to jack up
the lower bound on vector-space (which, for consistency, we'd want to
do also on a bunch of other diagrams packages).  For now, the current
code should still compile with older versions of vector-space.