split out path-drawing into separate function so clip can use it
Previously, the Renderable instance for Path R2 did not do any filling
or stroking; now it does. Clipping used to work by calling the
Renderable Path instance and then doing a C.clip.  But now this was
causing the clipping paths to actually get stroked and filled.  The
solution is that there is now a separate function which draws a Path
R2 without stroking or filling it, and both clipping and the
Renderable instance call it; clipping then does a C.clip and the
Renderable instance then does stroking and filling.